Fix o-o.myaddr.l.google.com DNSSource returning incorrect address
This DNSSource was configured with None server, thus using the default
recursive nameserver as the requester.
This caused the DNSSource to return the address of the recursive nameserver,
thus incorrect responses when what we wanted is to get our address.

Going up from subdomains, I found the the first subdomain with NS records is
l.google.com, whose NS are ns1.google.com, ns2.google.com, ns3.google.com, and
ns2.google.com.

Adding ns1.google.com as the server for this DNSSource solves the problem.