Fix DNSSource failing due to IP version mismatch
DNSSource failed when the IP version of the external IP resolution nameserver
did not match the first returned record type.
Ensuring the resolver configuration ip_strategy matches with the
requested family and record type solved the bug.

Additionally, added a fail-fast return when the requested family does not match
the record type.