DNS source: Make server non-optional and add constructor
By design, resolving the external address from a DNS source always
requires a server (using a recursive name server would return the
address of the recursive name server, which would only be correct if the
name server is behind the same NAT cone as the machine requesting the
external address).
This is solved by making the server parameter non-optional.

Additionally, add a public constructor so clients of the library can
use a non-hardcoded service.